requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
amqp==5.0.6
anyjson==0.3.3
asgi-redis==1.4.3
asgiref==3.2.10
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==20.2.0
Authlib==0.15.4
autobahn==20.7.1
Automat==20.2.0
billiard==3.6.4.0
blinker==1.4
cachetools==4.2.2
celery==5.0.0
certifi==2021.5.30
cffi==1.14.3
channels==3.0.3
channels-redis==3.1.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.3
ConfigArgParse==1.5
configobj==5.0.6
constantly==15.1.0
cryptography==3.1.1
daphne==3.0.2
defusedxml==0.6.0
distro==1.5.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.4.0
django-crontab==0.7.1
django-debug-toolbar==3.2.2
django-extensions==3.0.4
django-notifs==2.6.3
django-redis==5.0.0
django-rest-auth==0.9.5
django-salesforce-oauth==0.3.4
djangorestframework==3.11.1
djangorestframework-simplejwt==4.7.1
docker==5.0.0
drf-cached-instances==0.3.4
english==2020.7.0
eventlet==0.17.4
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==20.9.0
geventhttpclient==1.4.4
google-api-core==1.30.0
google-api-python-client==2.8.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==0.4.17
gunicorn==20.0.4
h11==0.10.0
hiredis==1.1.0
httplib2==0.19.1
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.0.0
incremental==17.5.0
Jinja2==3.0.1
jsonfield==3.1.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
kombu==5.1.0
locust==1.5.3
MarkupSafe==2.0.1
monotonic==1.6
msgpack==1.0.0
msgpack-python==0.5.6
multidict==5.1.0
neobolt==1.7.17
neotime==1.7.4
netifaces==0.10.4
numpy==1.19.1
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pandas==1.1.0
pansi==2020.7.3
pika==1.1.0
Pillow==7.2.0
prompt-toolkit==3.0.7
protobuf==3.17.3
psutil==5.7.2
py2neo==2021.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.3.1
PyHamcrest==2.0.2
PyJWT==2.1.0
pymongo==3.11.0
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-debian==0.1.32
python-http-client==3.3.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
pyzmq==22.1.0
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.7.2
SecretStorage==2.3.1
service-identity==18.1.0
simple-salesforce==1.11.4
six==1.15.0
slack==0.0.2
slackclient==2.9.3
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
ssh-import-id==5.11
tinydb==4.4.0
twilio==6.45.4
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
uvicorn==0.12.0
vine==5.0.0
wcwidth==0.2.5
websocket==0.2.1
websocket-client==1.1.0
websocket-client-py3==0.15.0
websockets==9.1
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.2.0
zope.event==4.5.0
zope.interface==5.1.0
